CHANGES: snmplib: PATCHES: 2780411: Fix compiler warnings
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@17639 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_parse_args.c
@@ -356,7 +356,7 @@ snmp_parse_args(int argc,
             break;
 
         case 't':
-            session->timeout = atof(optarg) * 1000000L;
+            session->timeout = (long)(atof(optarg) * 1000000L);
             if (session->timeout <= 0) {
                 fprintf(stderr, "Invalid timeout in seconds after -t flag.\n");
                 return (-1);
